requirements.txt
# Core
Flask>=2.0
python-dotenv>=0.19

# Database
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.1
psycopg2-binary>=2.9 # For PostgreSQL
# If using SQLite for testing/dev:
# sqlalchemy>=1.4

# Authentication
Flask-Login>=0.6
Werkzeug>=2.0 # Usually installed with Flask, explicit for hashing

# Optional: API Framework
# Flask-RESTful>=0.3

# Optional: Background Tasks
# Celery>=5.2
# redis>=4.0

# Optional: Data Analysis (add as needed)
numpy>=1.21
# pandas>=1.3
scikit-learn>=1.0